Sprint 1/ProximaCentauriMachine/AbdulAhad/requirements.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0
Npm install –g aws-cdk